Handle time value overflow in sqlite busy handler

+++ src/database-sqlite3.h
@@ -64,7 +64,7 @@ private:
 	sqlite3_stmt *m_stmt_begin;
 	sqlite3_stmt *m_stmt_end;
 
-	u32 busy_handler_data[2];
+	s64 busy_handler_data[2];
 };
 
 #endif
